Refactor mute-user→user-mute
