--- apps/web/src/components/domains/user/mute-user-list-item/index.ts
@@ -1,1 +1,0 @@
-export * from './mute-user-list-item';
