--- apps/web/src/components/domains/user/mute-user-list/index.ts
@@ -1,1 +1,0 @@
-export * from './mute-user-list';
