+++ apps/web/src/components/domains/user/user-mute-list-item/index.ts
@@ -1,0 +1,1 @@
+export * from './user-mute-list-item';
